Backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=1.10.0
pydantic-settings>=2.0.0  # Add this line
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.0
email-validator>=2.0.0
pyodbc>=4.0.39